refactor(sink): impl RowEncoder for JsonEncoder
